MySQL lastInsertId should be represented as Long type instead of Integer

+++ vertx-mysql-client/src/main/java/io/vertx/mysqlclient/MySQLClient.java
@@ -11,5 +11,5 @@ public interface MySQLClient {
   /**
    * SqlResult Property for last_insert_id
    */
-  PropertyKind<Integer> LAST_INSERTED_ID = () -> Integer.class;
+  PropertyKind<Long> LAST_INSERTED_ID = () -> Long.class;
 }
